disable Matomo cookies via _includes/head.html + pins ETT version
The head.html file was copied from the current ETT version (3.2.0). Hopefully, ETT will adopt the PR to disable Matomo cookies (ELIXIR-Belgium/elixir-toolkit-theme#314). When they do, the head.html file in this repo can be removed again and we can simply configure the disabling of Matomo cookies via _config.yml.

This commit also pins the ETT version to 3.2.0 to make sure we update it manually later.