Update authors of the task page
We realised that we have not clarified well what does it mean to become a page author. Everyone who does a PR or conttributes in some other way is a contributors. Page auhtors are contributors who write task and other pages. Doing a minor edit on a page makes someone a contributor but not a page author. Will have to apologise to Graeme for this and clarify things in the future.