add a very rough first cycle
Ripped the code directly from the RDMkit and replaced some of the words with those in the Research Software Lifecycle diagram.

The order doesn't make perfect sense this way, but for now let's just keep it simple to try out things. We can always make it into a 3D cylinder or cone to include tiers later.