link checker improvements
- Remove html and rst files from the checker (we are only interested in markdown files)
- Exclude private links (localhost, etc.)
- Add fallback extensions option. When the link checker finds a local link (a link to a page of RSQKit) without an explicit extension (.md or .html) it will try to find the file with the fallback extension as well. Set to .md now (by default no fallback extensions are set). This fixes a few of the local links.

Unfortunately, the last point still does not fix all local links, it only fixes those that lead to pages generated by markdown files located in the same directory. The problem is that while the markdown files are located in subdirectories, Jekyll build transforms all pages to html files in the root directory, without any subdirectory structure. This means that before building, the links in the markdown files cannot be properly resolved outside of their own directories, because the location of pages in other directories is unknown.