link checker: ignore local links
As explained in the previous commit message, most local links in markdown files cannot properly be resolved, so we skip them until we find a solution.

Possible solutions may include doing a build step before checking links and then running the link checker on the built HTML files. Disadvantage is that this introduces many autogenerated links (e.g. navigation links) that is wasteful to check. It also makes link checking much slower, because it has to wait for the build to succeed.